Software RAID: Identify the root fs via its UUID from image metadata
In order to install the bootloader on instances with software RAID,
the IPA needs to chroot into the deployed image's root fs. Currently,
the IPA assumes the first partition to contain the root fs, so this
breaks with images where this is not the case. The proposed change
extracts the root fs UUID from the image's metadata instead and
passes it to the IPA (in the same way this is done for partition
images). It keeps the current behaviour to send the UUID extracted
from the driver internal info in case the metadata is not set on
the image. For the time being, the IPA completely ignores the UUID
passed anyway, so to make this fully work the IPA will also need
to be patched.

Change-Id: I69bd53aa24b7b10384dc1399a0acaa21d38d6bd9
Story: #2006649
Task: #37081