Support burning configdrive into boot ISO
Adds the ability to write configdrive contents on the boot
images that ironic produces. Such boot image will be labeled
as `config-2` to facilitate tools like `cloud-init` or `Glean`
to find configuration source there.

As of this commit, the new code is not used by anything in
ironic yet.

Change-Id: Ib5ed0ce09ce97ddb914fa4f830265dfc24a0bc1a
Story: 2006691
Task: 36990