ironic-tempest-functional-python3 unused variables
The difference between the jobs `ironic-tempest-functional-python2`
and `ironic-tempest-functional-python3` should be only the python
version in use (for deployment and to run the functional tests).

Removing the variables `IRONIC_TEMPEST_WHOLE_DISK_IMAGE` and
`IRONIC_VM_EPHEMERAL_DISK`doesn't matter for the functional tests,
and they are set with different values in the python2 job.

Change-Id: I49ff241793f42de0f0cc9d98ae1f99940e581dcb