Remove the [pxe]ipxe_enabled configuration option
The ipxe_enabled configuration option was deprecated for quite a while,
this patch removes it. However the code impact by simply removing the
configuration option is quite large, so the strategy here is to register
the option dynamically but sticks to False, which essentially disabled
the ipxe support in the pxe interface.

Story: 2007003
Task: 37779

Change-Id: I499e627f780b577e351fd39be5fa93a34d384e64