Enable agent_token for virtual media boot
This provides an embedded agent_token through out of band
means which ensures greater security for such deployments.

Also some changes to the redfish virtual media boot unit
tests as the entirety of conductor utilties was mocked.

Downside with the pregenerated token, the logic had to be
slightly revised as the existence of a token with an older
client is a case that virtual media deployments will hit.

This was addressed, and tests updated as a result.

Story: 2007025
Task: 37819
Change-Id: Iaa2a52c2e53534cbf6998ad24f1f1184c0f222d8