Extend service timeout
Glance seems to be having issues getting started in time.

At least a couple times recently devstack exited because glance
was not running with-in 60 seconds. Logs in one case seemed to
lean to well over a minute, as such lets extend the timeout
and see if that stabilizes CI further.

[0] https://storage.bhs.cloud.ovh.net/v1/AUTH_dcaab5e32b234d56b626f72581e3644c/zuul_opendev_logs_b45/697584/10/check/ironic-tempest-ipa-wholedisk-bios-agent_ipmitool-tinyipa-indirect/b456a36/

Change-Id: I7d4457a456dedead4ad0d6a444be23a901b42111