Fix fast_track + agent_url update fix
In the case of fast_track, we do not power off agents
which allows for operators to perform a rapid transition,
such as those desired when performing a batch installation
of machines.

However, a security fix I1a5e3c2b34d45c06fb74e82d0f30735ce9041914
removed prior traces of an agent_url upon exiting cleaning.

So what was occuring:
- Agent was completing cleaning
- Conductor was removing 'agent_url' from driver internal info
- Agent was heartbeating, but no new agent_url was being recorded
  because how do we know if that is actually the agent.
- Deployment was started.
- Deployment fails upon next heartbeat as there is no 'agent_url'.

This patch places a conditional around the removal of the agent_url
value at the end of cleaning which allows nodes to transition as
through fast track as desired.

Change-Id: I2b5130a0845b327c2244d2d30a19842c46e1eed3
Story: 2007080
Task: 37991