Fix API docs for target_power_state response
The API docs stated that the values of target_power_state in a
Node State Summary are the same as the valid input values in a Change
Node Power State request (i.e. they include "rebooting", "soft
rebooting", and "soft power off"). However, in actuality these are all
compressed to either "power on" or "power off" by the
_calculate_target_state() function.

Change-Id: I8c8b8c707102275a48233fc853acfb65f556946e
Story: #2007089
Task: 38115