Clean up api controller base classes
The parent class of APIBase is no longer wsme.types.Base. Instead
there is a new ironic.api.controllers.base.Base as the parent. This
new Base class keeps the __init__() from wsme.types.Base, but not the
__registry__ registration.

As far as I can tell[1], this type registry is used to allow the
wsproperty datatype value to be a string, but all of our uses of
wsproperty use the real type types.uuid, so this registry is just
overhead.

The other changes are for some existing classes to extend the new Base
class instead of APIBase or wsme.types.Base. APIBase is now used only
by classes which represent real database objects, which is the only
situation where having a created_at, updated_at makes sense.

DeployStepType is excluded from this change as it will require extra
work to change its parent class.

Story: 1651346
[1] https://opendev.org/x/wsme/src/branch/master/wsme/types.py#L507

Change-Id: Ie687c270ed13b99486496a84df34e5973af1b9cd