Fix jsonpatch related tests
Latest release of jsonpatch fixed an issue with array boundaries[1]
but broke tests since JsonPatchConflict is now correctly raised
when a patch tries to replace an item that doesn't exist.
This patch fixes the issue with the tests including
JsonPatchConflict in the json exceptions intercepted by the method
apply_jsonpatch.

[1] https://github.com/stefankoegl/python-json-patch/commit/b3726f3a8bdcdf0f0841e078228014de8477b0ec

Change-Id: I6edf6eb7ae1e9f3aa7bc3220cd943a4849f4997c