tell reno to ignore the kilo branch
When reno 3.x runs under setuptools, it scans all of the branches it
can find, including any that look like they're closed and have an -eol
tag. The old kilo branch in this repository has a jumbled history that
somehow makes it look like it should include tags that it doesn't. We
know that there are no release notes in that branch, because reno
wasn't adopted while it was open. The releasenotes/source/index.rst
links to separate release notes in the wiki.

This patch tells reno to ignore that branch so that it doesn't throw
an exception when it gets confused about the old tag.

Story: #2007274
Task: #38710

Change-Id: I8047d805b0a3da4051a6d618ed9ac40875b8636a
Signed-off-by: Doug Hellmann <doug@doughellmann.com>
(cherry picked from commit 0f4b9e38a2924a693e17d815b16f19223f16632a)