Project Contributing updates for Goal
This change covers changes to meet the goal of establishing contributor
documentation, which Ironic largely already had, but required some
enhancement and clarification.

This also includes rough context of the PTL duties and revision to the
primary repository CONTRIBUTING document, as suggested in some of the
goal related discussions.

Change-Id: Ia47eb56d7eb3c19c99fa7a61fb5605037dd9ebee
Story: #2007236
Task: #38528