Change force_raw_images to use sha256 if md5 is selected
In order to enable ironic's conductor to execute on nodes
set for FIPS 140-2 compliance, we need to not explicitly
choose MD5.

In the case of forcing images to raw, we were calculating
the checksum at least once, if not twice. Now we will honor
the original algorithm unless it is MD5, at which point we
will default to SHA3-256, and only recalculate the checksum
once.

Change-Id: I408a2e461bebf1f6d9fa3e350eb7ab1a3544adad
Story: 2007306
Task: 38791