Make reservation checks caseless
If a conductor hostname is changed while reservations are
issued to a conductor with one hostname, such as 'hostName'
and then the process is restarted with 'hostname', then the
queries would not match the node and effectively the nodes
would become inaccessible until the reservation is cleared.

This patch changes the queries to use a case insenitive
like search, better known as ilike.

Special thanks to the nova team for finding a bug in the hash
ring calculation in the ironic virt driver where a customer
changed their nova-compute hostname, which inspired this
patch.

Change-Id: Ib7da925ba5ca6a82ba542e0f4ae4cf7f0d070835