Extend power sync timeout for Ericsson SDI
Ericsson SDI uses cached power state data in what is returned
to the API user, and typically this takes longer than 30 seconds
to provide an updated value to the API user.

As such, if we do not extend the timeout value, such users of
Ironic's Redfish API interface will have deployments fail as
the power state timeout will be encountered.

Change-Id: I0aa5131504b60b13d43c73c9a3be1f50f7855cbc