Implement the bios-interface for idrac-wsman driver
The BIOSInterface implementation for idrac-wsman Hardware Type was
missing and VendorInterface methods were used to configure BIOS instead.
This was not desired since it exposes a lot of vendor intricacies to the
operator and makes the existing mechanism unstable across Ironic releases.

This change introduces the implementation of BIOSInterface with the following
code changes:

1. update IDRACHardware class with supported_bios_interfaces() method
2. update setup.cfg with an entry for idrac-wsman in bios interfaces
3. Implement the DracWSManBIOS class with appropriate methods
4. Unit tests for the work

Change-Id: Ie85e30e2dfdfb1f199ac4d25f165e0cca033a4df
Co-authored-by: Richard Pioso <richard.pioso@dell.com>
Story: 2007400
Task: 38998