Fix issue where server fails to reboot
The iDRAC will return a "The command failed to set RequestedState" error
if the server is powered off and a reboot is requested.  In this
situation, convert the requested reboot into a power on to avoid this
error.

To minimize the chance of a race condition, it is critical to do this
check immediately before sending the power state change command.  This
keeps the window during which the server could change power states
without us knowing about it as small as possible.

Note that this error is returned in other scenarios as well.

Change-Id: I4c95959090423ff1929700f0fce596510447ac1c
Story: 2007487
Task: 39206