Add network_type to port local_link_connection
Add network_type to the port objects local_link_connection field.
The network_type can be set to either managed or unmanaged. When
the type is unmanaged other fields are not required. Use
unmanaged when the neutron network_interface is required, but
the network is in fact a flat network where no actual switch
management is done.

Story: 2007315
Task: 39161
Change-Id: I00c5ea22a8163c27c9ce4470e3713c282d0eb131