Replace oslo_utils.netutils type compares with ipaddress
We used netutils earlier on to have a backportable change
however the longer term goal was to replace the change with
using the python native ipaddress module directly.

For the cases where we can change IP version type compares,
we change them with this change.

Note: other uses of netutils still exist, and we should
eventually see if we can phase them out, however the remaining
uses are around MAC address validations.

Change-Id: I44336423194eed99f026c44b6390030a94ed0522