Release reservation when stoping the ironic-conductor service
If a conductor hostname is changed while reservations are
issued to a conductor with one hostname, such as 'hostname'
and then the process is restarted with 'new_hostname', then the
queries would not match the node and effectively the nodes
would become inaccessible until the reservation is cleared.

This patch clears the reservation when stoping the
ironic-conductor service to avoid the nodes becoming inaccessible.

Ref to: https://review.opendev.org/#/c/711765/

Change-Id: Id31cd30564ff26df0bbe4976ffe3f268b0dd3d7b