Extend timeout on CI job with automated cleaning
The ironic-tempest-ipa-partition-pxe_ipmitool CI job
has automated cleaning enable which introduces another boot
cycle where IPA starts and performs an operation. Due to the
way tempest is designed with the timeouts, it appears that
we are encountering one of the timeouts sporatically enough
to cause a high gate failure rate.

Extends the general timeout for the aformentioned test as
the ramdisk takes 400+ seconds to now boot.

Change-Id: I35fcf0d1f9340b0e906ea993c1b745850a4e3208