Don't break UEFI install with older IPAs
For Software RAID, we need to pass the boot mode to the IPA via a new
parameter. In case the IPA does not understand this new parameter yet,
we catch the error and fail for UEFI. This patch proposes to only fail
in the case of software RAID, since the older IPA can handle the general
case of UEFIi, e.g. with a whole disk image.

Change-Id: Id6b51dd19e83d20dc8d6d312cbec12d09bfd76c1