Stop configuring install_command in tox.
Currently, we are overriding 'install_command' to use 'pip'. This is
considered poor behavior and 'python -m pip' should be used instead:

https://snarky.ca/why-you-should-use-python-m-pip/

It turns out that this is the the default value provided by tox:

https://tox.readthedocs.io/en/latest/config.html#conf-install_command

So we can remove the line and simply use the default value.

openstack-discuss thread
http://lists.openstack.org/pipermail/openstack-discuss/2020-April/014237.html

Bumping  in requirements:
-alembic to 0.9.6
-keystoneauth1 to 3.18

Bumping libs in  in l-c:
-alembic to 0.9.6
-keystoneauth1 to 3.18
-pika to 0.10.0 (necessary for installation)

Change-Id: I059e841ab891ba0451d0423065e9727c8c77f337