Fix agent_client handling of embedded errors
The agent_client may get a general "command_error" field
returned to it upon commands and the agent may not properly
sinal that the command failed because we are reliant upon
the same data elsewhere in the ironic/agent interaction.

This resulted in the case where the embedded error signaled that
the error was method compatability as opposed to the actual method
command error.

This could cause higher level failures and prevent fallback logic
from detecting that we could try a different command.

We now consider the error type, and raise the appropriate exception
to signal that the issue may be an API compatability issue.

Change-Id: Ia2f63bd853632e1d7138901cf23fde1e261fc4d6