redfish: split reboot into power off followed by power on
The current implementation may result in a false positive when
the reboot request is ignored completely or when a node stays
powered on some time after the request is received.

Also make error messages a bit more useful by mentioning the desired
power state and avoiding redundant "Redfish".

Task: 39679
Story: 2007527
Change-Id: I35984e315948fbbcf216c40b38397953fb9dc699