New configuration parameter to use ipmitool retries
Add a new ``[ipmi]use_ipmitool_retries`` option. When set to
``True`` and timing is supported by ipmitool, the number of
retries and command interval will be passed to ipmitool so
that ipmitool will do the retries.  When set to ``False``,
ironic will do the retries.

The default is ``True``, so this will not change the current
behaviour which is to have ipmitool do the retries when
timing is supported.

Setting to ``False`` will help with certain BMCs which do
not support the Cipher Suites command.  In this case ipmitool
can take up to 10 seconds for each retry which results in a
total time exceeding ``[ipmi]command_retry_timeout``.

Change-Id: I1d0194e7c7ae9fcdd4665e6115ee26d10b14e480
Story: 2007632
Task: 39676
(cherry picked from commit 6520b373f456a6392e1b0ee6fdfaa632ce0d0dc3)