Increase callback timeout
One of the false positives I've observed in the last
24 hours is a job where the ramdisk did not DHCP on it's
network interfaces until the 550 second of runtime mark.

With a callback wait of 600 seconds, this means that the
deployment test job was failed moments later before the
ramdisk was fully online.

As such, lets move the timeout out a little as we're using
larger ramdisks now, and we need to allow for a longer window
at this point.

Change-Id: I26bd8f2f56732be1a75518950385caeb7b439b91
(cherry picked from commit 5f24029c40688594b7f116d3e3cf23b0b48bb446)