Extend vmedia job timeout
The vmedia job has an extra image mastering step which
generates about an additional copy of the kernel+ramdisk on
disk in the form of an ISO file.

This extra step, does add time to the overall build time, and
even with the current 700 second default, the kernel runtime on
the ramdisk was observed at 640 seconds when the machine was
powered off just after completing the heartbeat lookup with
ironic.

Change-Id: I90f9ab6aceffcb17e0c1ead059e10a68e4871fea