Fix fast track when exiting cleaning
When exiting cleaning, previously the agent token was purged
from ironic's database and agents continuing to run would not
be able to heartbeat to the conductor. With agent token, this
would orphan the agent such that it thought it had an agent
token, yet the conductor did not.

Change-Id: Id6f8609bcda369649d0f677aceed26ed5e72a313
(cherry picked from commit 5b785569b635a3d23509b009c4f575d1800c6f02)