Add virtualbmc as an extra dependency
The virtualbmc package is only needed in CI runs, so it doesn't quite
fit as a declared "normal" dependency of the project. But requirements
tooling only looks in [test-]requirements.txt and setup.cfg when
checking if a package is being used, so since it has not been declared
in any of these so far, it gets flagged as an unused dependency.

This is a workaround to fix this issue. It likely will never be
installed through this extra requirement path, but it will ensure the
package gets recognized as something being used by our projects.

Change-Id: Id7eb44954308b10d6097ee9894d58867fde6804b
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>