ISO ramdisk virtual media test enablement
Sets the settings to enable the ramdisk iso booting tests
including a bootable ISO image that should boot the machine.

NB: The first depends-on is only for temporary testing of another
which changes the substrate ramdisk interface. Since this change pulls
in tempest testing for iso ramdisk and uses it, might as well
use it to test if the change works or not as the other two patches
below are known to be in a good state.

Change-Id: I5d4213b0ba4f7884fb542e7d6680f95fc94e112e