Stop allocating double space for raw images
There are two possibilities when ironic conductor will be converting
instance image to raw format:

1. when iscsi deploy is used and [DEFAULT]force_raw_images config option
   is set to True;
2. when direct deploy is used, [agent]image_download_source config
   option is set to "http" and [DEFAULT]force_raw_images and
   [agent]stream_raw_images config options are set to True.

If one of the above conditions is met, ironic tries to allocate the disk
space needed for conversion even if the image is already raw, basically
doubling the disk space requirements for each image. This change adds a
check that skips the unnecessary conductor disk space allocation attempt
if the image is already raw.

Story: 2007830
Task: 40112
Change-Id: I730f9ec89ff1598d42777e4a503f46316eb85cad