Convert root controller to plain controller
The RestController._route method[1] contains a lot of logic which
exists to support implementing CRUD APIs, none of which is required to
server / and /v1.

This change converts the RestController based root controller with a
simple object based controller. This results in a shorter code path
for every API request.

The path / is handled by the index method, with the *args length check
returning a 404 to be consistent with the previous behaviour of
returning a 404 for a request like //v1

The _route method is replaced with a _lookup[2] method to handle paths
which are missing the v1 prefix (and to curate the trailing slash from
the path, which was been done by the RestController).

[1] https://github.com/pecan/pecan/blob/master/pecan/rest.py#L103

Change-Id: Ia962f9bf53af12fb308e3d66eff114b302638714
Story: 1651346
Task: 10551